requirements_dev.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.4.0
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
constantly==15.1.0
cryptography==41.0.1
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
isodate==0.6.1
logzero==1.7.0
multidict==6.0.5
numpy==2.1.0
pandas==2.2.2
pycparser==2.21
pycrypto==2.6.1
PyJWT==2.9.0
pyotp==2.8.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-telegram-bot==20.0
pytz==2023.3
rdflib==6.3.2
rdflib-jsonld==0.6.2
requests==2.31.0
rfc3986==1.5.0
setuptools==59.6.0
simplejson==3.19.1
six==1.16.0
smartapi-python==1.4.8
sniffio==1.3.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2024.1
urllib3==2.0.3
websocket-client==1.6.0
yarl==1.9.4
zope.interface==6.0
